Many small fixes
